Add macros and a whole lot more
Add DEFMACRO and MACROEXPAND.
Add NULL?, which tests if an object is nil.
Add MAP and LIST* to prelude.
Add +, -, *, / for any number of arguments.
Add visp.load_file to load and evaluate a file in an environment.
Add visp.load_prelude to load `prelude.visp' and evaluate it in an
environment.